src/pyproject.toml
[project]
name = "Bridge-X-RAG"
version = "0.1.0"
description = "Bridge-X RAG application"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=24.1.0",
    "cryptography>=44.0.2",
    "fastapi[standard]>=0.115.8",
    "langchain>=0.3.19",
    "langchain-community>=0.3.18",
    "motor>=3.7.0",
    "odmantic>=1.0.2",
    "passlib>=1.7.4",
    "pydantic-mongo>=2.3.0",
    "pydantic-settings>=2.8.0",
    "pyjwt>=2.10.1",
    "pymupdf>=1.25.3",
]
